fix: add helper functions for dealing with AnyXmlValue and export them
