+++ packages/helper/src/mosModel/index.ts
@@ -6,7 +6,8 @@ export * from './profile2'
 export * from './profile3'
 export * from './profile4'
 export * from './parseMosTypes'
-export { ensureArray, literal, omitUndefined, flattenXMLText } from './lib'
+export { literal, omitUndefined, flattenXMLText } from './lib'
+export * from '../utils/ensureMethods'
 export * from './ParseError'
 
 import { AnyXMLObject } from '@mos-connection/model'
